backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
bcrypt==4.0.1
openai==1.55.3
mcp==1.1.2
pydantic-ai==0.0.10
